alt som har med analytics kan leve i umami.ts for nå
